+++ front/src/modules/companies/components/companyBoardOptions.tsx
@@ -1,7 +1,7 @@
 import { CompanyBoardCard } from '@/companies/components/CompanyBoardCard';
 import { NewCompanyProgressButton } from '@/companies/components/NewCompanyProgressButton';
 
-import { BoardOptions } from '../types/BoardOptions';
+import { BoardOptions } from '../../pipeline-progress/types/BoardOptions';
 
 export const companyBoardOptions: BoardOptions = {
   newCardComponent: <NewCompanyProgressButton />,
