--- front/src/modules/pipeline-progress/states/NewBoardCardContext.ts
@@ -1,3 +1,0 @@
-import { createContext } from 'react';
-
-export const NewBoardCardContext = createContext<string | null>(null);
