--- front/src/modules/pipeline-progress/states/boardItemsState.ts
@@ -1,8 +1,0 @@
-import { atom } from 'recoil';
-
-import { EntityProgressDict } from '../components/EntityProgressBoard';
-
-export const boardItemsState = atom<EntityProgressDict>({
-  key: 'boardItemsState',
-  default: {},
-});
