--- front/src/modules/pipeline-progress/states/selectedBoardItemsState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const selectedBoardItemsState = atom<string[]>({
-  key: 'selectedBoardItemsState',
-  default: [],
-});
